requirements.txt
pydantic>=2.7.0
openai>=1.3.5
pyyaml>=6.0.1
python-dotenv>=1.0.0
flask>=2.3.3
